2 commit

+++ src/main/java/com/jareer/lms/app/dtos/UniversityDTO.java
@@ -1,5 +1,5 @@
 package com.jareer.lms.app.dtos;
 
 
-public record UniversityDTO(String name, String address, short openYear) {
+public record UniversityDTO(String name, String address, Integer openYear) {
 }
